Refactored Validators and sql converters module
