+++ src/yads/converters/__init__.py
@@ -1,3 +1,4 @@
-from .sql import SparkSQLConverter, SQLConverter, SQLGlotConverter
+from .sql.sql_converter import SQLConverter, SparkSQLConverter
+from .sql.ast_converter import SQLGlotConverter
 
 __all__ = ["SQLConverter", "SparkSQLConverter", "SQLGlotConverter"]
