+++ src/yads/registries/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseRegistry  # type: ignore[reportMissingImports]
+
+__all__ = ["BaseRegistry"]
