--- src/yads/registry/__init__.py
@@ -1,3 +1,0 @@
-from .base import BaseRegistry
-
-__all__ = ["BaseRegistry"]
